Split up pydocstyle rules
