+++ src/rules/pydocstyle/rules/capitalized.rs
@@ -1,0 +1,36 @@
+use crate::ast::types::Range;
+use crate::checkers::ast::Checker;
+use crate::docstrings::definition::{DefinitionKind, Docstring};
+use crate::registry::Diagnostic;
+use crate::violations;
+
+/// D403
+pub fn capitalized(checker: &mut Checker, docstring: &Docstring) {
+    if !matches!(docstring.kind, DefinitionKind::Function(_)) {
+        return;
+    }
+
+    let body = docstring.body;
+
+    let Some(first_word) = body.split(' ').next() else {
+        return
+    };
+    if first_word == first_word.to_uppercase() {
+        return;
+    }
+    for char in first_word.chars() {
+        if !char.is_ascii_alphabetic() && char != '\'' {
+            return;
+        }
+    }
+    let Some(first_char) = first_word.chars().next() else {
+        return;
+    };
+    if first_char.is_uppercase() {
+        return;
+    };
+    checker.diagnostics.push(Diagnostic::new(
+        violations::FirstLineCapitalized,
+        Range::from_located(docstring.expr),
+    ));
+}
